refactor: 게시글 상세 조회 수정, 삭제 버튼 컴포넌트 분리
